+++ geolocator_android/android/src/main/java/com/baseflow/geolocator/location/NotificationImportance.java
@@ -1,0 +1,7 @@
+package com.baseflow.geolocator.location;
+
+public enum NotificationImportance {
+    Default,
+    High,
+    Max,
+}