+++ geolocator_apple/apple/Classes/Handlers/GeolocationHandler.h
@@ -18,8 +18,9 @@ typedef void (^GeolocatorResult)(CLLocation *_Nullable location);
 
 - (CLLocation *_Nullable)getLastKnownPosition;
 
-- (void)requestPosition:(GeolocatorResult _Nonnull)resultHandler
-           errorHandler:(GeolocatorError _Nonnull)errorHandler;
+- (void)requestPositionWithDesiredAccuracy:(CLLocationAccuracy)desiredAccuracy
+                             resultHandler:(GeolocatorResult _Nonnull)resultHandler
+                              errorHandler:(GeolocatorError _Nonnull)errorHandler;
 
 - (void)startListeningWithDesiredAccuracy:(CLLocationAccuracy)desiredAccuracy
                            distanceFilter:(CLLocationDistance)distanceFilter