+++ geolocator_apple/macos/Classes/geolocator-umbrella.h
@@ -1,0 +1,1 @@
+../../apple/Classes/geolocator-umbrella.h