+++ geolocator_android/CHANGELOG.md
@@ -1,3 +1,11 @@
+## 3.1.0
+
+- Adds support to make a foreground service and continue processing location updates when the application is moved into the background.
+
+## 3.0.4
+
+- Fixes Android embedding v2 warning when compiling the example App.
+
 ## 3.0.3
 
 - Added a default `intervalDuration` value of 5000ms to prevent the `getCurrentPosition` method to return a cached Location.
@@ -57,4 +65,3 @@
 ## 1.0.0
 
 - Initial open source release.
-